Fixes MonoBehaviourBaseEditor
Marks the object as dirty after setting the property. When included in a
instantiated prefab, modifying the property does not cause the field to
be marked for serialization, causing it to lose the modified value after
play.
Removes equality check. The property itself should do the check if
necessary, by default it should be possible to reassign the same value
and cause the setter to be called. The check was also not working with
value types.